backend/requirements.txt
# Backend API Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# OpenAI (optional fallback)
openai>=1.12.0

# Google Gemini AI (primary)
google-generativeai>=0.8.0

# Vector Database
qdrant-client>=1.7.0

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.0

# Text Processing
langchain-text-splitters>=0.0.1

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
